# _Reversa_MasterSaaS/openapi/mastersaas-api.yaml
openapi: "3.1.0"
info:
  title: MasterSaaS API
  version: "1.0.0"
  description: |
    Especificação OpenAPI gerada pelo Reversa v1.2.14 a partir de engenharia reversa do frontend.
    O sistema atual opera em mock in-memory — esta spec define o contrato de backend a implementar.
    Autenticação delegada ao Supabase Auth; endpoints usam Bearer JWT emitido pelo Supabase.
servers:
  - url: https://api.mastersaas.com/v1
    description: Produção
  - url: http://localhost:3000/v1
    description: Local
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido pelo Supabase Auth (anon key ou service_role key)
  schemas:
...
    TutorialCategory:
      type: string
      enum: [getting-started, first-sale, scaling-sales, campaigns]
    Tutorial:
      type: object
      required: [id, youtubeUrl, category, ctaTo, order, active, required]
      properties:
        id:
          type: string
        youtubeUrl:
          type: string
          format: uri
        youtubeId:
          type: string
          readOnly: true
          description: "Extraído de youtubeUrl — suporta watch?v=, youtu.be/, embed/, shorts/"
        category:
          $ref: "#/components/schemas/TutorialCategory"
        title:
          type: string
        description:
          type: string
          nullable: true
        ctaTo:
          type: string
          description: "Rota interna de CTA (ex: /products)"
        order:
          type: integer
        active:
          type: boolean
        required:
          type: boolean

    # ─── WhatsApp Integration ─────────────────────────────────────────────────
...
paths:
...
  /tutorials:
    get:
      tags: [Tutorials]
      summary: Lista tutoriais disponíveis
      parameters:
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/TutorialCategory"
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Tutoriais ordenados por `order` ASC
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tutorial"

    post:
      tags: [Tutorials, Admin]
      summary: Cria ou atualiza tutorial (upsert por id — admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tutorial"
      responses:
        "200":
          description: Tutorial criado ou atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tutorial"

  # ── WhatsApp Integration ──────────────────────────────────────────────────
...